// micro-apps/spinal-board-commissioning/src/interfaces/IGetAllBuildingsRes.ts
/*
 * Copyright 2022 SpinalCom - www.spinalcom.com
 *
 * This file is part of SpinalCore.
 *
 * Please read all of the following terms and conditions
 * of the Free Software license Agreement ("Agreement")
 * carefully.
 *
 * This Agreement is a legally binding contract between
 * the Licensee (as defined below) and SpinalCom that
 * sets forth the terms and conditions that govern your
 * use of the Program. By installing and/or using the
 * Program, you agree to abide by all the terms and
 * conditions stated or referenced herein.
 *
 * If you do not agree to abide by these terms and
 * conditions, do not demonstrate your acceptance and do
 * not install or use the Program.
 * You should have received a copy of the license along
 * with this file. If not, see
 * <http://resources.spinalcom.com/licenses.pdf>.
 */

export interface IGetAllBuildingsRes {
  name: string;
  aliasName: string;
  urlBos: string;
  apiUrl: string;
  clientId: string;
  clientSecret: string;
  address: string;
  description: string;
  picture_base64: string;
  location: ILocation;
  type: string;
  id: string;
  directModificationDate: number;
  details: IDetails;
  [key: string]: any;
}
interface ILocation {
  lat: string;
  lng: string;
}
interface IDetails {
  area: number;
}


export interface IContext {
  dynamicId: number;
  staticId: string;
  name: string;
  type: string;
}

export  interface ICategory {
  dynamicId: number;
  staticId: string;
  name: string;
  type: string;
  icon: string;
}

export interface IGroup{
  dynamicId: number;
  staticId: string;
  name: string;
  type?: string
}


export interface ItemInGroup {
  dynamicId: number;
  staticId: string;
  name: string;
  type: string;
  bimFileId: string;
  dbid: number;
}


export interface ItemPositon {
  
    dynamicId: 0;
    staticId: string;
    name: string;
    type: string;
    info: {
      context: {
        dynamicId: 0;
        staticId: string;
        name: string;
        type: string
      },
      building: {
        dynamicId: 0;
        staticId: string;
        name: string;
        type: string;
        adress: string;
        area: 0
      },
      floor: {
        dynamicId: 0;
        staticId: string;
        name: string;
        type: string;
      },
      room: {
        dynamicId: 0;
        staticId: string;
        name: string;
        type: string
      }
    }
  }